auto-caption-video/requirements.txt
Pillow>=10.0.0
edge-tts>=6.1.0
